# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithDetails"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment (status, notes)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, doctor, patient]
        phone:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        qualifications:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        consultationFee:
          type: ["number", "null"]
        bio:
          type: ["string", "null"]
        availableDays:
          type: ["string", "null"]
        availableHours:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserWithProfile"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, doctor, patient]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
    UserWithProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            doctorProfile:
              $ref: "#/components/schemas/DoctorProfile"
            patientProfile:
              $ref: "#/components/schemas/PatientProfile"
...
    DoctorProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        specialty:
          type: string
        qualifications:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        consultationFee:
          type: ["number", "null"]
        bio:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        availableDays:
          type: ["string", "null"]
        availableHours:
          type: ["string", "null"]
      required:
        - id
        - userId
        - specialty
    DoctorWithUser:
      allOf:
        - $ref: "#/components/schemas/DoctorProfile"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
          required:
            - user
...
    PatientProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        medicalHistory:
          type: ["string", "null"]
      required:
        - id
        - userId
    PatientWithUser:
      allOf:
        - $ref: "#/components/schemas/PatientProfile"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
          required:
            - user
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - doctorId
        - scheduledAt
        - status
        - createdAt
    AppointmentWithDetails:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            doctor:
              $ref: "#/components/schemas/DoctorWithUser"
            patient:
              $ref: "#/components/schemas/PatientWithUser"
    CreateAppointmentBody:
      type: object
      properties:
        doctorId:
          type: integer
        patientId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
          format: date-time
        reason:
          type: ["string", "null"]
      required:
        - doctorId
        - scheduledAt
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes:
          type: ["string", "null"]
        scheduledAt:
          type: string
          format: date-time
...